DERBY-428: PreparedStatement.executeBatch hangs if batch is too large
This patch contains a server-side change, a client-side change, and a test.

The server-side change is to call ensureLength() in DDMWriter.startDDM().
The DDMWriter working buffer is designed to dynamically grow to accomodate
the data being written; this dynamic growth is implemented using a coding
rule which requires that all DDMWriter internal routines must call
ensureLength to communicate the buffer size requirements prior to writing
bytes into the buffer. StartDDM was missing the call to ensureLength.

The client-side change is due to the fact that DRDA imposes a hard limit
of 65535 elements in a single correlated request because the correlation
identifier is a two byte unsigned integer. Prior to this change, the
correlation identifiers would wrap around when we added the 65536th
element into the batch, which breaks the DRDA protocol rule that requires
correlation IDs in a single request be always increasing. This change
causes the client to throw an exception if it is asked to execute a batch
containing more than 65534 elements. The reason for the number 65534,
rather than 65535, is that the value 0xFFFF is reserved for special use.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@387895 13f79535-47bb-0310-9956-ffa450edef68