DERBY-1136: getFloat throws LossOfPrecisionConversionException on MAX_VALUE
This patch contains a regression test for DERBY-1136. The regression test
verifies that we can successfully retrieve a value of Float.MAX_VALUE
from the database using ResultSet.getFloat.

The test runs for both the embedded driver and the DerbyNetClient driver.
The test is dummied out for the JCC driver, since that driver still suffers
from this bug (but we want to share the same set of master files).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@389575 13f79535-47bb-0310-9956-ffa450edef68