DERBY-349: Re-enable parameterMapping batch tests for DerbyNetClient
Due to bug DERBY-349, the parameterMapping tests of various data types
in the executeBatch() configuration had been disabled, because those
tests were hanging. The tests are no longer hanging, so this submission
re-enables the tests.

This change also modifies the parameterMapping test program so that it
knows how to unwind the BatchUpdateException when looking for an
underlying Invalid Conversion exception. This means that we can have
simpler and easier-to-read master output files, which makes the test
easier to maintain.

This change also deletes the separate jdk14 master file, since it is
identical to the primary master file.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@391086 13f79535-47bb-0310-9956-ffa450edef68