DERBY-1076: Resolve the output differences in upgrade test
Attaching a patch which updates the master file for upgrade test. The
master file is updated to reflect changes in the metadata test. Also
there were extra lines in the master file in the previous checkin. I
have also replaced 10.1.2.2 with 10.1.2.1 as Knut had suggested. This
version info will hopefully go away with my next patch to allow
running the test as part of harness. Still there are few diffs because
of the diffs in the generated SQL id which look like this line:

,APP,REFACTION1,A,,APP,REFACTNONE,A,1,3,3,SQL060329093642910,SQL060329093642450,7 

This will also go away once the test can be run with the harness,
which filters out the generated ids.

Patch contributed by Deepa Remesh <dremesh@gmail.com>.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@391255 13f79535-47bb-0310-9956-ffa450edef68