DERBY-1144 With client PooledConnection.getConnection() does not reset holdability and isolation state properly
This patch addresses this issue by changing PooledConnection.getConnection() to reset the connection to the datasource defaults.

Code changes:
Just a small code change in ClientPooledConnection to pass true to the reset method to reset the defaults.

This change has the side effect of making DERBY- 1173 harder to reproduce. But I don't think it fixes that bug. With the change, the checkDataSource test hangs only very intermittently. The fixer of that bug will have to go back to a previous version to get a easily reproducable case.


Tests -
Changed the checkDataSource30.java test to test holdability,isolation, and autocommit reset.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@391369 13f79535-47bb-0310-9956-ffa450edef68