DERBY-1152: Failures in sysinfo tests with different classpath settings
Depending on the way the classpath was set, the sysinfo_withproperties and
sysinfo tests were unreliable. The problem is that the underlying sysinfo
classes are present in your classpath in several possible ways: you might
be pointing directly at the classes directory, or you might be pointing to
one of: derby.jar, derbytools.jar, or derbynet.jar, all three of which have
the sysinfo classes in them. This means that there are four possible
sections of the derby_tests.policy file that might be in effect for the
sysinfo classes during a test run, and those four sections were inconsistent.

This change modifies the derby_tests.policy file so that, at least for the
classpath and locale file permissions that are used by sysinfo, the policy
file expresses a consistent set of permissions for each of the four code
sources. The change also updates the master files to reflect the chosen
permissions, which are to allow access to the classpath and the locale files.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@391387 13f79535-47bb-0310-9956-ffa450edef68