DERBY-514 Integrate upgrade tests into test suite. Contributed by Deepa Remesh
Attaching a patch 'derby-514-buildfiles-v1.diff' which adds the build files for upgrade tests. This patch also adds the properties file for upgrade test which specifies not to run with security manager as more work is needed in this area (DERBY-1135).

This patch affects upgrade tests only. This is an incremental patch and does not resolve the issue fully. Please take a look at this patch. Thanks.

 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@391806 13f79535-47bb-0310-9956-ffa450edef68