DERBY-514 Integrate upgrade tests into test suite Contributed by Deepa Remesh.
Attaching a patch 'derby-514-patch2-runtest-v1.diff' which enables the upgrade test to run with the test harness.

Summary:

* findCodeBase method in harness/jvm.java is changed to public. This method is used by upgrade test to get the location of new jar files.

* Adds the other derby jars to the jar file list in UpgradeTester. This will allow the test to run in client framework. I tried running the test in client framework and it looks like this will need new master file and some more work.

* In UpgradeTester, File.toURL method is used when creating class loader. This seems to be a better way to construct the URL.

* Master file update

This patch combined with the previous patch (derby-514-buildfiles-v1.diff) will allow the upgrade test to be run using RunTest. The location of old jars has to be passed in as a property in jvmflags. Command to run:
java -Djvmflags=-DderbyTesting.oldJarLocation=<location of 10.1 jars> org.apache.derbyTesting.functionTests.harness.RunTest upgradeTests/Upgrade_10_1_10_2.java 
 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@391844 13f79535-47bb-0310-9956-ffa450edef68