DERBY-1189
inplace compress could self deadlock on a latch/latch conflict.  Changed
the insert page find code for defragment to never return a page which is
higher than the source page of the row.  Previously the upper level code
made this decision, but that was too late.  As part of the change also 
optimized the "unfilled" search to be biased towards those pages near
the front of the container.  Checkin includes simple sql script which 
reproduced the deadlock prior to the fix.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@392194 13f79535-47bb-0310-9956-ffa450edef68