DERBY-1184: registerOutParameter(int,int,String) should throw exception
Patch contributed by Kristian Waagan (Kristian.Waagan@Sun.com)

The method 'CallableStatement.registerOutParameter(int,int,String)'
does nothing in the client driver. As stated in DERBY-447, the method throws
a not-implemented exception in the embedded driver. The method should be
changed to do this on the client side as well.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@392304 13f79535-47bb-0310-9956-ffa450edef68