DERBY-1189
backporting fix (svn 392194) from trunk to 10.1 branch:

inplace compress could self deadlock on a latch/latch conflict.  Changed
the insert page find code for defragment to never return a page which is
higher than the source page of the row.  Previously the upper level code
made this decision, but that was too late.  As part of the change also
optimized the "unfilled" search to be biased towards those pages near
the front of the container.  Checkin includes simple sql script which
reproduced the deadlock prior to the fix.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@393855 13f79535-47bb-0310-9956-ffa450edef68