DERBY-842 : Internationalize messages in PreparedStatement.java to Section.java in client/am.
This passes derbyall on JDK14.  I also verified that the tests whose
master files affected in the JDK14 run were also run on JDK13 and JDK16
and fixed output files for those platforms as well, as needed.

I also removed a few spurious SQL States based on the test I ran to find
SQL States with no messages.  More of those to come in a separate patch.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@393967 13f79535-47bb-0310-9956-ffa450edef68