DERBY-1187
backported fix for DERBY-1187 from trunk to 10.1 branch, targeted for the
10.1.3 release.  Original comments from fix (392439):
Defragment pass was not being committed properly, so when rows at
front of table were being deleted and freeing pages the defragment
pass did not move the rows to the free pages.  Longer term the
defragment pass should commit more often, see DERBY-1188.
Added extra test to OnlineCompressTest.java for this case.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@398761 13f79535-47bb-0310-9956-ffa450edef68