DERBY-1229: sysinfo and sysinfo_withProperties tests fail with SecurityException
This change modifies the handling of security exceptions in the Sysinfo tool.
If a security exception occurs while processing an element in the classpath,
sysinfo will now format that security exception's message into the output,
and then will continue on to process the other elements in the classpath.

This ensures that sysinfo provides more detailed and more complete information
about the classpath information in the environment.

Also, the sed properties for the sysinfo tests are now updated to mask out
the getProtectionDomain security exception, since that is an expected
exception in these tests (in fact, there are many expected security exceptions
in these tests).



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@400366 13f79535-47bb-0310-9956-ffa450edef68