DERBY-900 -  Remove use of String(byte[]) and String(byte[], int, int) constructors in network client leading to non-portable behaviour
Decided earlier to break up the derby900.p2.diff.txt, so I am attaching now a patch with the cleanup changes separately in derby900_p2_cleanup.diff.txt. Changes are:

-- remove unnecessary conversion from sqlstate string to bytes and back to string , in parseSQLDCGRP(Sqlca[] rowsetSqlca, int lastRow) in NetConnectionReply.
This has been done by adding a constructor in NetSqlca to take in the sqlstate as string.
-- cleanup method getSqlState() in Sqlca.
-- cleanup of ccsid_ in Sqlca as it is not used.

Please note, the parseSQLDCGRP method in the diff shows more changes because of change in indentation because of removal of the try block.

-- ran derbynetclientmats on ibm142/linux OK.

Contributed by Sunitha Kambhampati




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@406120 13f79535-47bb-0310-9956-ffa450edef68