DERBY-1234: Verify that we raise SQLException when calling methods on closed java.sql objects
Final patch for DERBY-1234. The patch makes the methods in the client
driver check whether the object is closed. With this patch,
jdbcapi/StmtCloseFunTest.java has the same behaviour with the client
driver as with the embedded driver so the DerbyNetClient canon could
be removed. Also, jdbc4/ClosedObjectTest.junit runs cleanly both on
embedded and on the client, and it has been added to the jdbc40 suite.

In addition to adding checkForClosedXXX() calls, this patch

  a) implements Connection.getTypeMap() with the correct generic
     signature in NetConnection40 and LogicalConnection40 to avoid
     compiler warnings

  b) implements isWrapperFor() (with check for closed object) in
     Statement, so that it could be removed from Statement40,
     PreparedStatement40 and CallableStatement40


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@406280 13f79535-47bb-0310-9956-ffa450edef68