DERBY-1219: jdbcapi/checkDataSource test hangs intermittently with client
This patch contributed by Deepa Remesh (dremesh@gmail.com).

This patch enables the checkDataSource and checkDataSource30 tests to
run with the client framework by removing the code which shuts down
the system in the middle of the test. This is the code which causes
the intermediate hang.

The hanging problem has been logged as a separate issue, DERBY-1326.

The shutdown is a valuable part of the test because it verifies that the
global transaction state is valid even after the database has been shut
down, so once the hang problem has been resolved, this test should be
modified again to re-enable the shutdown processing with the client framework.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@406776 13f79535-47bb-0310-9956-ffa450edef68