DERBY-1156: Committing this for Suresh, reviewed by Mike Matrigali.
Here is Mike's review comments:
I have reviewed this patch and run a full set of tests on windows XP / sun jdk1.4.2 which passed. I think this patch should be committed as is. It is a good incremental checkin for this feature. It provides a set of tools for the rest of the project, and gets the code to the point that it can encrypt an existing unencrypted db. Future checkins should address a couple of things:
1) obviously more testing. Some quick notes, the current tests may want to show that accessing the newly encrypted db with a bad password does not work - just to insure all the url;s are not just being totally ignored. Also want to test abort - probably need a testing codepoint to cause the failure at the right time.
2) fix up the comments, there are some typo's .
3) usual nit - some greater than 80 char lines.

I would have committed this change now, but svn still is not back.

Patch submitted by Suresh Thalamati (suresh.thalamati@gmail.com)



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@407366 13f79535-47bb-0310-9956-ffa450edef68