DERBY-1338: Client tests fail with NoClassDefFound: DRDAProtocolExceptionInfo
Patch contributed by Dag Wanvik (dag.wanvik@sun.com)

Work around a classloader bug involving interrupt handling during class
loading. If the first request to load the DRDAProtocolExceptionInfo class
occurs during shutdown, the loading of the class may be aborted when the
Network Server calls Thread.interrupt() on the DRDAConnThread.  By including
a static reference to the DRDAProtocolExceptionInfo class here, we ensure
that it is loaded as soon as the DRDAConnThread class is loaded,
and therefore we know we won't be trying to load the class during shutdown.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@416012 13f79535-47bb-0310-9956-ffa450edef68