DERBY-1454: DRDA Protocol Exception when DSS is exactly 32767 bytes
The server is disconnecting the client connection. Client is sending a CNTQRY,
the server is receiving it and then server disconnects the client. In doneData
and writeFDODTA, it looks like if (writer.getDSSLength() >= blksize) ,
then the splitQRYDTA is called. From testing, it seems like they are hitting
an edge case, where the writer.getDSSLength() is equal to blksize and this
DSS is the only dss in the send buffer. 

The patch corrects the two splitQRYDTA calls to use > rather
than >= to determine if the DSS needs to be split.

The patch also contains a new test case in derbynet/prepStmt.java
to demonstrate the bug and its fix.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@416959 13f79535-47bb-0310-9956-ffa450edef68