DERBY-1455: Intermittent failure in derbynet/runtimeinfo.java
The problem occurs when the test has opened four connections, closes
three of them and runs getRuntimeInfo() on the server. The test
expects that there is one active session on the server, but in some
cases, the network server has not had enough time to clean up after
the closed connections, and more sessions are listed as active.

The patch solves this timing issue by invoking Thread.sleep(1000)
before calling getRuntimeInfo().


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@417131 13f79535-47bb-0310-9956-ffa450edef68