DERBY-937, committing change by      Andreas Korneliussen.
notes by Andreas:
I was able to reproduce this problem on every run on a very fast laptop (it was not reproducible on any other Windows lab machine I have tried). This laptop happens to be the same kind as Ole uses for the nightly tests.

Adding a 10 second sleeping period after the population of the tables, did not have any effect. I therefore tried to do run a compress on the tables (based on the assumption that statistics is updated on compress), and now the test does not fail for me anymore.

Attached is the patch which makes this test stop failing. The patch does not seem to have any sideeffects on other platforms (Solaris) tested, however the test will use more time. 

note by mike:
I ran this change 20 times and it worked on my laptop, it would always fail
a few times out of 20 before the fix.  I am not sure why this fix works as
the test seems to already create indexes after tables so statistics should
already be up to date. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@418108 13f79535-47bb-0310-9956-ffa450edef68