DERBY-578 - Grouped select from temporary table raises null pointer exception in byte code generator DERBY-1464 - runtimestatistics can show that an index is being used even when it isn't
Contributed by Manish Khettry

The problem is simple enough-- we didn't have a conglomerate name for temporary tables. I fixed the code to behave more like what fillInScanArgs does.

Earlier, we would set the indexName field in DistinctScanResult to the conglomerate name (cd.getName()) used to scan the table. If the conglomerate was the base table itself then this was just plain wrong. The change, for this patch, passes null if no index is being used.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@418672 13f79535-47bb-0310-9956-ffa450edef68