DERBY-1417: Add new, lengthless overloads to the streaming api
'derby-1417-2a-rstest-refactor.diff' refactors and rewrites parts of
jdbc4/ResultSetTest.junit. A few utility methods have been introduced,
and a TestSetup-class (anonymous) was used to create the table
once. This was done because the previous implementation could cause
other tests to fail if one test failed.

The utility methods follow a pattern that will also be used by a
number of tests written for the new lengthless overloads.

Patch contributed by Kristian Waagan.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@420497 13f79535-47bb-0310-9956-ffa450edef68