DERBY-1481: Client driver: ResultSet.beforeFirst() gives protocol error on scrollable, updatable result sets that are downgraded to read-only
When the result set is downgraded from updatable to read-only because
the query generating the result set cannot produce an updatable result
set, the result set will be downgraded on the server side, and a
warning (SQLCARD) indicating the downgrade will be sent to the
client. Warnings on the server side are not cleared after they are
sent to the client causing the server to send the same warning several
times.

Positioning commands like ResultSet.beforeFirst() that do not return
any data also do not expect warnings to be returned. The protocol
error was being caused by this downgrade warning that was being sent
several times, one of them being in a response to a
ResultSet.beforeFirst() command.

The attached patch (derby-1481.diff) fixed the problem by clearing the
warnings after they are sent so that the same warning will not be sent
more than once.

The patch was contributed by Fernanda Pizzorno.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@421887 13f79535-47bb-0310-9956-ffa450edef68